requirements.txt
bandit==1.9.2
blinker==1.9.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
filelock==3.20.2
Flask==3.1.2
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
license-expression==30.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
packageurl-python==0.17.6
packaging==25.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.1
pluggy==1.6.0
py-serializable==2.1.0
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
sortedcontainers==2.4.0
stevedore==5.6.0
tomli==2.3.0
tomli_w==1.2.0
urllib3==2.6.2
Werkzeug==3.1.4
